Move postprocess to transform
